requirements.txt
fastapi==0.95.2
requests==2.30.0
pydantic==1.10.7
uvicorn[standard]==0.22.0
pandas==1.5.3
fastapi_utils==0.2.1
google-cloud-bigquery==3.10.0
google-cloud-storage==2.5.0
google-cloud-bigquery-storage==2.19.1
db-dtypes==1.0.4
pycountry==22.3.5
inflect==6.0.0
rocketry==2.5.1
google-cloud-translate==3.9.0